requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-cors-headers==2.4.0
djangorestframework==3.9.4
future==0.17.1
idna==2.8
mysqlclient==1.3.13
PyJWT==1.7.1
pytz==2018.7
qiniu==7.2.2
requests==2.21.0
urllib3==1.24.1
